+++ app/src/device.h
@@ -5,7 +5,7 @@
 
 #include "config.h"
 #include "common.h"
-#include "net.h"
+#include "util/net.h"
 
 #define DEVICE_NAME_FIELD_LENGTH 64
