+++ app/src/stream.c
@@ -10,12 +10,12 @@
 
 #include "config.h"
 #include "compat.h"
-#include "buffer_util.h"
 #include "decoder.h"
 #include "events.h"
-#include "lock_util.h"
-#include "log.h"
 #include "recorder.h"
+#include "util/buffer_util.h"
+#include "util/lock.h"
+#include "util/log.h"
 
 #define BUFSIZE 0x10000
