+++ app/src/sys/unix/net.c
@@ -1,4 +1,4 @@
-#include "net.h"
+#include "util/net.h"
 
 #include <unistd.h>
