+++ app/src/util/str_util.c
